For the final project submission, we assumed that we have to create our own repo and add our project files to it by inviting you the repo. However, now it has come to light that we had to fork the ‘final_project_2020Sp' repo and submit all our necessary files to it. I had created the repo ‘https://github.com/akashsukhija/MonteCarloFinalProject' and we have committed all our code and changes to it since the past few days. I have already sent you the invite a few minutes ago.
For now, only I am committing to the ‘final_project_2020Sp' repo all our changes would be visible in the repo link given earlier.

Please let us know if this will pose a problem. Be assured that we have been working diligently on implementing our code and all our work is visible in the repo.

I apologize for this confusion.